Refactor and version fix
